fix more unit tests

+++ tests/test_automl.py
@@ -15,8 +15,8 @@ def generate_data_csv_file():
         ROOT_PATH, 'profile_data/profile_data_for_test_automl.csv')
     destination_file = os.path.join(
         ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv')
-
-    shutil.copyfile(source_file, destination_file)
+    if not os.path.exists(destination_file):
+        shutil.copyfile(source_file, destination_file)
 
 
 def delete_data_file(path):
@@ -41,8 +41,8 @@ def test_automl_bs_find_pruning_ratio1():
         model, hyperparameters, None, None, 20, writing_info=writing_info)
     assert pruning_ratio == pytest.approx(
         0.9), "test_automl_bs_find_pruning_ratio failed"
-    delete_data_file(os.path.join(
-        ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
+    # delete_data_file(os.path.join(
+    #     ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
 
 
 @pytest.mark.timeout(1)
@@ -63,8 +63,8 @@ def test_automl_bs_find_pruning_ratio2():
     pruning_ratio, accuracy, latency = automl_bs_find_pruning_ratio(
         model, hyperparameters, None, None, 99, writing_info=writing_info)
     assert pruning_ratio == None, "test_automl_bs_find_pruning_ratio failed"
-    delete_data_file(os.path.join(
-        ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
+    # delete_data_file(os.path.join(
+    #     ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
 
 
 @pytest.mark.timeout(1)
@@ -81,5 +81,5 @@ def test_automl_bs_find_pruning_ratio3():
         model, hyperparameters, None, None, 85, writing_info=writing_info)
     assert pruning_ratio == pytest.approx(
         0.4), "test_automl_bs_find_pruning_ratio failed"
-    delete_data_file(os.path.join(
-        ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
+    # delete_data_file(os.path.join(
+    #     ROOT_PATH, 'profile_data/profile_data_resnet18_noise_0.0_poison_0.0.csv'))
